resmon: Test SFD and SFDF registers
Verify that FDB counter is increased/decreased according to SFD EMADs.

Allocate two entries with the same FID and verify that they are both
flushed when SFDF EMAD with the corresponding flush type is received.

Add a separate file to handle SFD and SFDF EMADs and import it to
resmon-test.sh

Signed-off-by: Amit Cohen <amcohen@nvidia.com>