resmon: resmon-stat: Split function for future use
resmon_stat_lh_update_nostats() searches a key in hash-table and only if
the key doesn't exist, insert it to the table.

For SFD register, which will be added in the next patch, two hash-tables
will be used and in some cases only lookup or only insert is needed.

Split the function to resmon_stat_lh_lookup_entry() and
resmon_stat_lh_insert(), so the each one can be used for SFD tables also.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>